Move definitions into the internal package
