+++ balloon/src/main/kotlin/com/skydoves/balloon/extensions/TextViewExtension.kt
@@ -25,8 +25,8 @@ import androidx.appcompat.content.res.AppCompatResources
 import androidx.core.text.HtmlCompat
 import com.skydoves.balloon.IconForm
 import com.skydoves.balloon.IconGravity
-import com.skydoves.balloon.NO_INT_VALUE
 import com.skydoves.balloon.TextForm
+import com.skydoves.balloon.internal.NO_INT_VALUE
 import com.skydoves.balloon.vectortext.VectorTextView
 import com.skydoves.balloon.vectortext.VectorTextViewParams
